+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * new methods for `Args` (*the returned object from* `Console.get_args()`):<br>
   - the `len()` function can now be used on `Args` (*the returned object from* `Console.get_args()`)
   - the `Args` object now also has the dict like methods `keys()`, `values()` and `items()`
+  - you can also get the args as a dict with the `dict()` method
   - you can now use the `in` operator on `Args`
 * new methods for `ArgResult` (*a single arg-object from inside `Args`):<br>
   - you can now use the `bool()` function on `ArgResult` to directly see if the arg exists